requirements.txt
asttokens==3.0.0
blinker==1.9.0
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
dnspython==2.8.0
executing==2.2.1
Flask==3.1.2
Flask-PyMongo==3.0.1
ipykernel==6.30.1
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.5
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.3
python-dateutil==2.9.0.post0
pywin32==311
pyzmq==27.1.0
six==1.17.0
spyder-kernels==2.5.2
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.2.14
Werkzeug==3.1.3
